Table - Fix Preselected Row & Test Structure Refactor (#834)
