+++ packages/terra-list/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update table test links to be 'Table Tests' and 'Single Select Table Links'
 
 1.11.0 - (September 26, 2017)
 ------------------
