+++ packages/terra-site/src/TestLinks.jsx
@@ -45,6 +45,7 @@ const TestLinks = () => (
       <li><Link to="/tests/search-field-tests">Search Field</Link></li>
       <li><Link to="/tests/signature-tests">Signature Tests</Link></li>
       <li><Link to="/tests/single-select-list-tests">SingleSelectList Tests</Link></li>
+      <li><Link to="/tests/single-select-table-tests">SingleSelectTable Tests</Link></li>
       <li><Link to="/tests/slide-group-tests">SlideGroup Tests</Link></li>
       <li><Link to="/tests/slide-panel-tests">SlidePanel Tests</Link></li>
       <li><Link to="/tests/status-tests">Status Tests</Link></li>
