+++ packages/terra-table/src/SingleSelectableRows.jsx
@@ -29,9 +29,9 @@ class SingleSelectableRows extends React.Component {
     if (!rows || !rows.length) {
       return null;
     }
-    // Find the first row which is selected and is selectable
+    // Find the first row which is selected
     for (let i = 0; i < rows.length; i += 1) {
-      if (rows[i].props.isSelected && rows[i].props.isSelectable) {
+      if (rows[i].props.isSelected) {
         return i;
       }
     }
