+++ packages/terra-table/tests/nightwatch/single-select-table/components/SelectableTableNoRows.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Table from '../../../lib/Table';
+import Table from '../../../../lib/Table';
 
 const SingleRowSelectableTable = () => (
   <Table isStriped={false}>
