+++ packages/terra-table/tests/nightwatch/table/components/StripedTable.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import Table from '../../../lib/Table';
+import Table from '../../../../lib/Table';
 
-const TableWithHighlightedRows = () => (
-  <Table isStriped={false}>
+const StripedTable = () => (
+  <Table id="Table">
     <Table.Header>
       <Table.HeaderCell content={'Name'} key={'NAME'} minWidth={'small'} />
       <Table.HeaderCell content={'Address'} key={'ADDRESS'} minWidth={'medium'} />
       <Table.HeaderCell content={'Phone Number'} key={'PHONE_NUMBER'} minWidth={'large'} />
     </Table.Header>
     <Table.Rows>
-      <Table.Row isSelected key={'PERSON_0'}>
+      <Table.Row key={'PERSON_0'}>
         <Table.Cell content="John Smith" key="NAME" />
         <Table.Cell content="123 Adams Drive" key="ADDRESS" />
         <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
@@ -24,18 +24,8 @@ const TableWithHighlightedRows = () => (
         <Table.Cell content="213 Raymond Road" key="ADDRESS" />
         <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
       </Table.Row>
-      <Table.Row isSelected key={'PERSON_3'}>
-        <Table.Cell content="Mario Smith" key="NAME" />
-        <Table.Cell content="213 Road" key="ADDRESS" />
-        <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
-      </Table.Row>
-      <Table.Row key={'PERSON_4'}>
-        <Table.Cell content="Louie Smith" key="NAME" />
-        <Table.Cell content="213 Raymond  St" key="ADDRESS" />
-        <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
-      </Table.Row>
     </Table.Rows>
   </Table>
 );
 
-export default TableWithHighlightedRows;
+export default StripedTable;
