+++ packages/terra-table/tests/nightwatch/table/components/TableWithSortIndicator.jsx
@@ -1,14 +1,15 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
-import Table from '../../../lib/Table';
+import Table from '../../../../lib/Table';
 
-const StripedTable = () => (
-  <Table id="Table">
+const TableWithSortingIndicator = () => (
+  <Table isStriped={false}>
     <Table.Header>
       <Table.HeaderCell content={'Name'} key={'NAME'} minWidth={'small'} />
-      <Table.HeaderCell content={'Address'} key={'ADDRESS'} minWidth={'medium'} />
+      <Table.HeaderCell content={'Address'} key={'ADDRESS'} minWidth={'medium'} sort={'asc'} />
       <Table.HeaderCell content={'Phone Number'} key={'PHONE_NUMBER'} minWidth={'large'} />
     </Table.Header>
-    <Table.Rows>
+    <Table.SingleSelectableRows>
       <Table.Row key={'PERSON_0'}>
         <Table.Cell content="John Smith" key="NAME" />
         <Table.Cell content="123 Adams Drive" key="ADDRESS" />
@@ -24,8 +25,8 @@ const StripedTable = () => (
         <Table.Cell content="213 Raymond Road" key="ADDRESS" />
         <Table.Cell content="111-222-3333" key="PHONE_NUMBER" />
       </Table.Row>
-    </Table.Rows>
+    </Table.SingleSelectableRows>
   </Table>
 );
 
-export default StripedTable;
+export default TableWithSortingIndicator;
